Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -7,8 +7,7 @@ int main(){
     scanf("%d",&a);
     scanf("%d",&b);
     scanf("%d",&c);
-    int avg = (a+b+c)/3;
+    float avg = (a+b+c)/3;
     printf("Average: %.2f", avg);
     return 0;
-    
 }
